TL: update on run3D simulation function
